refactor(feature-currency): add global error handler
