+++ src/main/java/com/exchange/calculator/currency/gloabl/error/ControllerAdvice.java
@@ -1,6 +1,5 @@
-package com.exchange.calculator.currency.ui;
+package com.exchange.calculator.currency.gloabl.error;
 
-import com.exchange.calculator.currency.dto.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
